requirements.txt
Django==1.7
django-braces==1.4.0
django-registration-redux==1.1
six==1.8.0